+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.4.2] - 2019-09-06
+
 ### Added
 
 - Create roles and permissions for organizations and courses imported via the
@@ -82,7 +84,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Static and media files are stored in AWS S3 buckets and distributed _via_
   Amazon CloudFront
 
-[unreleased]: https://github.com/openfun/fun-mooc/compare/v0.4.1...HEAD
+[unreleased]: https://github.com/openfun/fun-mooc/compare/v0.4.2...HEAD
+[0.4.2]: https://github.com/openfun/fun-mooc/compare/v0.4.1...v0.4.2
 [0.4.1]: https://github.com/openfun/fun-mooc/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/openfun/fun-mooc/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/openfun/fun-mooc/compare/v0.2.0...v0.3.0